Refactor profile fetching
